requirements.txt
undetected-chromedriver>=3.5.0
Pillow>=9.0.0
beautifulsoup4>=4.9.3
requests>=2.26.0
Flask>=2.0.0
Flask-CORS>=3.0.10
stem>=1.8.0
bcrypt>=3.2.0 